refactor daemon packages
