+++ internal/daemon/common.go
