+++ internal/daemon/common_test.go
